+++ src/azure-cli/azure/cli/command_modules/profile/tests/latest/test_profile.py
@@ -27,16 +27,6 @@ class ProfileTest(ScenarioTest):
             assert result['tenantId'] == MOCKED_TENANT_ID
             assert result['user']['name'] == MOCKED_USER_NAME
 
-    def test_list_locations(self):
-        result = self.cmd('account list-locations').get_output_in_json()
-        extended_result = self.cmd('account list-locations --include-extended-locations').get_output_in_json()
-        assert isinstance(result, list)
-        assert len(extended_result) >= len(result)
-        # Verify there is an item with displayName 'East US'.
-        assert any('East US' == loc['displayName'] for loc in result)
-        assert any('geography' in loc['metadata'] for loc in result)
-        assert any('availabilityZoneMappings' in loc for loc in result)
-
 
 if __name__ == '__main__':
     unittest.main()
