+++ src/azure-cli/azure/cli/command_modules/resource/aaz/latest/account/_list_locations.py
