+++ src/azure-cli/azure/cli/command_modules/resource/aaz/profile_2018_03_01_hybrid/__init__.py
