+++ src/azure-cli/azure/cli/command_modules/resource/aaz/profile_2020_09_01_hybrid/account/__cmd_group.py
