+++ src/azure-cli/azure/cli/command_modules/resource/aaz/profile_2020_09_01_hybrid/account/_list_locations.py
