+++ src/azure-cli/azure/cli/command_modules/resource/linter_exclusions.yml
