+++ src/azure-cli/azure/cli/command_modules/resource/tests/latest/recordings/test_list_locations.yaml
@@ -11,7 +11,7 @@ interactions:
       Connection:
       - keep-alive
       User-Agent:
-      - AZURECLI/2.52.0 (AAZ) azsdk-python-core/1.26.0 Python/3.10.11 (Windows-10-10.0.22621-SP0)
+      - AZURECLI/2.52.0 (AAZ) azsdk-python-core/1.26.0 Python/3.10.9 (Windows-10-10.0.22621-SP0)
     method: GET
     uri: https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/locations?api-version=2022-12-01
   response:
@@ -556,17 +556,19 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Fri, 15 Sep 2023 04:21:57 GMT
+      - Mon, 18 Sep 2023 02:54:32 GMT
       expires:
       - '-1'
       pragma:
       - no-cache
       strict-transport-security:
       - max-age=31536000; includeSubDomains
-      vary:
-      - Accept-Encoding
+      x-cache:
+      - CONFIG_NOCACHE
       x-content-type-options:
       - nosniff
+      x-msedge-ref:
+      - 'Ref A: E0A28E0623B042799B83F98AC6778B75 Ref B: TYO201100116039 Ref C: 2023-09-18T02:54:30Z'
     status:
       code: 200
       message: OK
@@ -584,7 +586,7 @@ interactions:
       ParameterSetName:
       - --include-extended-locations
       User-Agent:
-      - AZURECLI/2.52.0 (AAZ) azsdk-python-core/1.26.0 Python/3.10.11 (Windows-10-10.0.22621-SP0)
+      - AZURECLI/2.52.0 (AAZ) azsdk-python-core/1.26.0 Python/3.10.9 (Windows-10-10.0.22621-SP0)
     method: GET
     uri: https://management.azure.com/subscriptions/00000000-0000-0000-0000-000000000000/locations?includeExtendedLocations=true&api-version=2022-12-01
   response:
@@ -1262,17 +1264,19 @@ interactions:
       content-type:
       - application/json; charset=utf-8
       date:
-      - Fri, 15 Sep 2023 04:21:59 GMT
+      - Mon, 18 Sep 2023 02:54:34 GMT
       expires:
       - '-1'
       pragma:
       - no-cache
       strict-transport-security:
       - max-age=31536000; includeSubDomains
-      vary:
-      - Accept-Encoding
+      x-cache:
+      - CONFIG_NOCACHE
       x-content-type-options:
       - nosniff
+      x-msedge-ref:
+      - 'Ref A: 8E5FB6BFB3E54F7FBA4168D16B27A203 Ref B: TYO201100117049 Ref C: 2023-09-18T02:54:33Z'
     status:
       code: 200
       message: OK
